+++ packages/qvism-preset/src/vars/color/manner-temp/l3.mjs
@@ -1,2 +1,2 @@
-export const text = "var(--seed-color-manner-temp-l3-text)";
 export const bg = "var(--seed-color-manner-temp-l3-bg)";
+export const text = "var(--seed-color-manner-temp-l3-text)";