+++ packages/qvism-preset/src/vars/color/manner-temp/l6.d.ts
@@ -1,2 +1,2 @@
-export declare const text = "var(--seed-color-manner-temp-l6-text)";
 export declare const bg = "var(--seed-color-manner-temp-l6-bg)";
+export declare const text = "var(--seed-color-manner-temp-l6-text)";